requirements.txt
slack-bolt==1.18.0
slack-sdk==3.27.0
google-generativeai==0.8.0
feedparser==6.0.11
httpx==0.27.0
supabase==2.9.1
python-dotenv==1.0.1
schedule==1.2.2
flask==3.0.0
